refactor: move http helpers to module
